Move bulk of the CoreCLR delegate implementation to Delegate.CoreCLR.cs
- Matches structure of NativeAOT delegate impletation
- Avoids Unsafe.As casts
- Avoids C# "new slots" that are breaking code produced by managed C++. Fixes app compat bug #130612.
