+++ src/libraries/System.Private.CoreLib/src/System/MulticastDelegate.cs
@@ -3,11 +3,14 @@
 
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 using System.Runtime.Serialization;
 
 namespace System
 {
 #pragma warning disable CS0660, CS0661 // Defining operators, but not overriding Equals/GetHashCode
+    [ClassInterface(ClassInterfaceType.None)]
+    [ComVisible(true)]
     public abstract partial class MulticastDelegate : Delegate, ISerializable
     {
         // This constructor is called from the class generated by the
